+++ packages/config/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./AgentForgeConfig.js";
+export * from "./loadConfig.js";